+++ contracts/LSP6KeyManager/LSP6Utils.sol
@@ -189,7 +189,7 @@ library LSP6Utils {
 
         keys[1] = LSP2Utils.generateArrayElementKeyAtIndex(
             _LSP6KEY_ADDRESSPERMISSIONS_ARRAY,
-            uint128(arrayLength)
+            arrayLength
         );
         values[1] = abi.encodePacked(_address);
